+++ api/feedback.ts
@@ -3,6 +3,12 @@ import { octokit } from '../server/auth'
 
 export default function Feedback(req: NowRequest, res: NowResponse) {
   if (req.method === 'POST') {
+    const userNavigator = Object.keys(req.body.navigator)
+      .map((navigatorKey) => {
+        return `${navigatorKey}: ${req.body.navigator[navigatorKey]}`
+      })
+      .join('\n')
+
     octokit
       .request('POST /repos/{owner}/{repo}/issues', {
         owner: 'evillt',
@@ -10,8 +16,16 @@ export default function Feedback(req: NowRequest, res: NowResponse) {
         title: `${req.body.emotion}${req.body.title}`,
         labels: ['user feedback'],
         body: `
-UserAgent: \`${req.body.userAgent}\`
-`,
+### Email
+
+${req.body.email}
+
+### UserNavigator
+
+\`\`\`
+${userNavigator}
+\`\`\`
+`.trim(),
       })
       .then(() => {
         res.status(200)