+++ src/app-entries/prefer-appreance.ts
@@ -1,8 +1,0 @@
-import { isDarkMode } from '@/utils'
-import { doHTMLDarkTheme } from '@/utils/preference/switchAppreance'
-
-if (isDarkMode) {
-  doHTMLDarkTheme('add')
-} else {
-  doHTMLDarkTheme('remove')
-}