+++ src/hooks/useUserWallets.ts
@@ -3,15 +3,7 @@ import { Wallet } from '@/model/Wallet'
 import { ref } from 'vue'
 
 export function useUserWallets() {
-  // FIXME set a non-empty array to prevent oldChildren empty
-  // https://github.com/vuejs/vue-next/issues/2804
-  const userWallets = ref<Wallet[]>([
-    {
-      name: '',
-      icon: '',
-      balance: null,
-    },
-  ])
+  const userWallets = ref<Wallet[]>([])
 
   readWallets().then((wallets) => {
     userWallets.value = wallets